JSW-50651, JSW-52093: lock all calls to esp_transport_close (#1)
A mostly minimal fix for the double free described in JSW-50651 and JSW-52093 is to guard calls to *esp_transport_close()*. This changeset does that, using the existing mutex in the websocket client that guards some access to the client and its transport.

In the crashes that have been reported, both threads call *esp_websocket_client_abort_connection()*, so a more minimal fix would be to only lock calls to that function. But there are a couple of other callers of *esp_transport_close()* that might also have this issue.

Alternately, a more maximal fix would be to lock all access to the client and its transport. But that has side effects, such as locking the client while *esp_websocket_client_task()* blocks up to a second on `esp_transport_poll_read(client->transport, 1000)`; so it'd require more analysis and testing.

Regardless of what we choose here, it may change when we upstream the fix, based on feedback from the esp-websocket-client maintainers.

NB: *esp_websocket_client_abort_connection()* is sometimes called from a caller that has already locked the mutex, but it's a recursive mutex (created using *xSemaphoreCreateRecursiveMutex()*), so that's ok.